// src/types/colors.ts
type AllowedColors =
  | "#FF6666"
  | "#66FF99"
  | "#FF99CC"
  | "#6699FF"
  | "#A0522D"
  | "#FAD02E"
  | "#c02942"
  | "#CC66FF";

// type AllowedColors =
//   | "coral-red"
//   | "mint-green"
//   | "bubblegum-pink"
//   | "sky-blue"
//   | "sienna-brown"
//   | "sunflower-yellow"
//   | "crimson-red"
//   | "lavender-purple";

export type { AllowedColors };